delete TaskDetailsPage and change logic to show task details on the TasksListPage, from there user can get to the task edit page
