+++ pages/_app.js
@@ -17,7 +17,7 @@ export default function App({ Component, pageProps }) {
       <GlobalStyle />
       <ToastContainer
         position="top-center"
-        autoClose={5000}
+        autoClose={3000}
         hideProgressBar={false}
         newestOnTop={false}
         closeOnClick
