+++ pages/edit/index.js
@@ -1,9 +1,9 @@
 import { useRouter } from "next/router";
 import useSWR from "swr";
 import styled from "styled-components";
-import RegularTaskInputForm from "../../../components/RegularTaskInputForm";
+import RegularTaskInputForm from "../../components/RegularTaskInputForm";
 import { toast } from "react-toastify";
-import Layout from "../../../components/Layout";
+import Layout from "../../components/Layout";
 
 const headerText = "edit task";
 const homeButtonShow = true;
@@ -14,16 +14,16 @@ const BannerMessageSaved = () => <div>Task saved!</div>;
 export default function TaskEditPage() {
   const router = useRouter();
   const { isReady } = router;
-  const { dynamicId } = router.query;
+  const { id } = router.query;
   const {
     data: existingTaskData,
     isLoading,
     error,
     mutate,
-  } = useSWR(`/api/tasks/${dynamicId}`);
+  } = useSWR(`/api/tasks/edit?id=${id}`);
 
   async function editTask(existingTaskData) {
-    const response = await fetch(`/api/tasks/${dynamicId}`, {
+    const response = await fetch(`/api/tasks/edit?id=${id}`, {
       method: "PATCH",
       headers: {
         "Content-Type": "application/json",
@@ -33,7 +33,7 @@ export default function TaskEditPage() {
     if (response.ok) {
       mutate();
     }
-    router.push(`/tasks/${dynamicId}`);
+    router.push(`/`);
 
     // Info banner
     toast.success(<BannerMessageSaved />, {
@@ -56,7 +56,7 @@ export default function TaskEditPage() {
         onSubmit={editTask}
         formName={"edit-task"}
         existingTaskData={existingTaskData}
-        backLink={dynamicId}
+        backLink={`/`}
       />
     </Layout>
   );
